# web/docs/swagger.yaml
definitions:
...
  models.ConsoleLog:
    properties:
      id:
        type: integer
      result_id:
        type: integer
      type:
        type: string
      value:
        type: string
    type: object
  models.Cookie:
    properties:
      domain:
        type: string
      expires:
        type: string
      http_only:
        type: boolean
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      priority:
        type: string
      result_id:
        type: integer
      secure:
        type: boolean
      session:
        type: boolean
      size:
        type: integer
      source_port:
        type: integer
      source_scheme:
        type: string
      value:
        type: string
    type: object
  models.Header:
    properties:
      id:
        type: integer
      key:
        type: string
      result_id:
        type: integer
      value:
        type: string
    type: object
  models.NetworkLog:
    properties:
      error:
        type: string
      id:
        type: integer
      mime_type:
        type: string
      remote_ip:
        type: string
      request_type:
        $ref: '#/definitions/models.RequestType'
      result_id:
        type: integer
      status_code:
        type: integer
      time:
        type: string
      url:
        type: string
    type: object
  models.RequestType:
    enum:
    - 0
    - 0
    type: integer
    x-enum-varnames:
    - HTTP
    - WS
  models.Result:
    properties:
      console:
        items:
          $ref: '#/definitions/models.ConsoleLog'
        type: array
      content_length:
        type: integer
      cookies:
        items:
          $ref: '#/definitions/models.Cookie'
        type: array
      failed:
        description: Failed flag set if the result should be considered failed
        type: boolean
      failed_reason:
        type: string
      file_name:
        description: Name of the screenshot file
        type: string
      final_url:
        type: string
      headers:
        items:
          $ref: '#/definitions/models.Header'
        type: array
      html:
        type: string
      id:
        type: integer
      is_pdf:
        type: boolean
      network:
        items:
          $ref: '#/definitions/models.NetworkLog'
        type: array
      perception_hash:
        type: string
      probed_at:
        type: string
      protocol:
        type: string
      response_code:
        type: integer
      response_reason:
        type: string
      technologies:
        items:
          $ref: '#/definitions/models.Technology'
        type: array
      title:
        type: string
      tls:
        $ref: '#/definitions/models.TLS'
      url:
        type: string
    type: object
  models.TLS:
    properties:
      cipher:
        type: string
      encrypted_client_hello:
        type: boolean
      id:
        type: integer
      issuer:
        type: string
      key_exchange:
        type: string
      protocol:
        type: string
      resultid:
        type: integer
      san_list:
        items:
          $ref: '#/definitions/models.TLSSanList'
        type: array
      server_signature_algorithm:
        type: integer
      subject_name:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
  models.TLSSanList:
    properties:
      id:
        type: integer
      tls_id:
        type: integer
      value:
        type: string
    type: object
  models.Technology:
    properties:
      id:
        type: integer
      result_id:
        type: integer
      value:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /results/detail/{id}:
    get:
      consumes:
      - application/json
      description: Get details for a result.
      parameters:
      - description: The screenshot ID to load.
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Result'
      summary: Results detail
      tags:
      - Results
...
swagger: "2.0"